:recycle: EmbeddingRepositoryの実装をインフラレイヤーに移動
