+++ src/infrastructures/external_services.rs
@@ -1,0 +1,1 @@
+pub(crate) mod openai_embedding_repository;
